backend/pyproject.toml
[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[project]
name = "promo-constructor-backend"
version = "0.1.0"
description = "FastAPI backend for Promo Constructor"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1,<1.18.0",
    "fastapi>=0.128.0,<0.129.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "psycopg[binary]>=3.2.12,<3.3.0",
    "PyJWT>=2.10.1,<3.0.0",
    "pwdlib[argon2]>=0.3.0,<0.4.0",
    "python-multipart>=0.0.20,<0.0.21",
    "SQLAlchemy>=2.0.44,<2.1.0",
    "uvicorn[standard]>=0.38.0,<0.39.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
